+++ Editor/Provider/SimulatorAdaptivePerformanceSubsystem.cs
@@ -500,7 +500,7 @@ namespace UnityEditor.AdaptivePerformance.Simulator.Editor
             public int MaxGpuPerformanceLevel { get { return 4; } }
 
             /// <summary>
-            /// Simulator subsystem try to initialize initializes successfully. 
+            /// Simulator subsystem try to initialize initializes successfully.
             /// </summary>
             /// <returns>true if initialized</returns>
             protected override bool TryInitialize()