+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.0.1] - 2023-04-25
+
+### Fixed
+- APB-19: Fixed a missing provider popup message appearing on platforms when not provider is available but Adaptive Performance is installed.
+
 ## [5.0.0] - 2023-04-10
 
 ### Added